+++ wayland-backend/src/lib.rs
@@ -24,6 +24,12 @@
 //!
 //! Both the `wayland-client` and `wayland-server` crates follow this principle, so everything will "Just
 //! Work" when using them.
+//!
+//! ## Logging
+//!
+//! This crate can generate some runtime error message (notably when a protocol error occurs). By default
+//! those messages are printed to stderr. If you activate the `log` cargo feature, they will instead be
+//! piped through the `log` crate.
 
 #![warn(missing_docs, missing_debug_implementations)]
 #![forbid(improper_ctypes, unsafe_op_in_unsafe_fn)]
@@ -43,6 +49,16 @@ macro_rules! message {
         }
     }
 }
+
+// internal imports for dispatching logging depending on the `log` feature
+#[cfg(feature = "log")]
+#[allow(unused_imports)]
+use log::{debug as log_debug, error as log_error, info as log_info, warn as log_warn};
+#[cfg(not(feature = "log"))]
+#[allow(unused_imports)]
+use std::{
+    eprintln as log_error, eprintln as log_warn, eprintln as log_info, eprintln as log_debug,
+};
 
 #[cfg(any(test, feature = "client_system", feature = "server_system"))]
 #[cfg_attr(docsrs, doc(cfg(any(feature = "client_system", feature = "server_system"))))]
@@ -73,6 +89,7 @@ mod types;
  * They'll be optimized out when unused.
  */
 
+#[cfg(feature = "log")]
 #[no_mangle]
 extern "C" fn wl_log_rust_logger_client(msg: *const std::os::raw::c_char) {
     let cstr = unsafe { std::ffi::CStr::from_ptr(msg) };
@@ -80,6 +97,7 @@ extern "C" fn wl_log_rust_logger_client(msg: *const std::os::raw::c_char) {
     log::error!("{}", text);
 }
 
+#[cfg(feature = "log")]
 #[no_mangle]
 extern "C" fn wl_log_rust_logger_server(msg: *const std::os::raw::c_char) {
     let cstr = unsafe { std::ffi::CStr::from_ptr(msg) };