+++ wayland-client/src/lib.rs
@@ -158,6 +158,12 @@
 //!   method (see `wayland-backend` docs), and then make it into a [`Connection`] using
 //!   [`Connection::from_backend()`]. Similarly, you can make [`ObjectId`]s from the `*mut wl_proxy` pointers
 //!   using `ObjectId::from_ptr()`, and then make the proxies using [`Proxy::from_id`].
+//!
+//! ## Logging
+//!
+//! This crate can generate some runtime error message (notably when a protocol error occurs). By default
+//! those messages are printed to stderr. If you activate the `log` cargo feature, they will instead be
+//! piped through the `log` crate.
 
 #![allow(clippy::needless_doctest_main)]
 #![warn(missing_docs, missing_debug_implementations)]
@@ -187,6 +193,16 @@ pub use wayland_backend::protocol::WEnum;
 
 pub use conn::{ConnectError, Connection};
 pub use event_queue::{Dispatch, EventQueue, QueueHandle, QueueProxyData};
+
+// internal imports for dispatching logging depending on the `log` feature
+#[cfg(feature = "log")]
+#[allow(unused_imports)]
+use log::{debug as log_debug, error as log_error, info as log_info, warn as log_warn};
+#[cfg(not(feature = "log"))]
+#[allow(unused_imports)]
+use std::{
+    eprintln as log_error, eprintln as log_warn, eprintln as log_info, eprintln as log_debug,
+};
 
 /// Generated protocol definitions
 ///